# Add WriteTo trait and implement it for Message struct