--- a/sdk/src/bytes_serializable.rs
+++ b/sdk/src/bytes_serializable.rs
@@ -11,4 +11,9 @@ pub trait BytesSerializable {
     fn from_bytes(bytes: Bytes) -> Result<Self, IggyError>
     where
         Self: Sized;
+
+    /// Computes the size of the struct in bytes.
+    fn size(&self) -> usize {
+        unimplemented!("size")
+    }
 }