--- a/sdk/src/lib.rs
+++ b/sdk/src/lib.rs
@@ -31,3 +31,4 @@ pub mod topics;
 pub mod users;
 pub mod utils;
 pub mod validatable;
+pub mod write_to;